Add files via upload
This animation uses the Fibonacci sequence of numbers to create arcs that spiral out to what seems like infinity. with this animation the apparent motion is that you are spiraling down into the hole. forever. there is rainbow effect given and the spiral is created multiple times next to itself to create the look of a surface and to give it the multicolored look. this animation does not take up too much memory because it is a static image being rotated. and the only thing changing is the colors of each line.